langid/requirements.txt
accelerate==0.31.0
datasets==2.20.0
evaluate==0.4.2
librosa==0.10.2.post1
matplotlib==3.7.1
numpy==1.25.2
openpyxl==3.1.4
pandas==2.0.3
pycm==4.0
python-iso639==2024.4.27
scikit-learn==1.2.2
seaborn==0.13.1
soundfile==0.12.1
transformers==4.41.2